+++ lib/src/attributes/dynamic/dark_mode.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:mix/src/helpers/extensions.dart';
 
 import '../base_attribute.dart';
 
@@ -11,6 +12,6 @@ class DarkModeAttribute extends DynamicAttribute {
   const DarkModeAttribute(Attribute attribute) : super(attribute);
   @override
   bool shouldApply(BuildContext context) {
-    return Theme.of(context).brightness == Brightness.dark;
+    return context.isDarkMode;
   }
 }