+++ lib/src/helpers/extensions.dart
@@ -1,5 +1,45 @@
 import 'package:flutter/material.dart';
 
+/// Possible screen sizes
+enum ScreenSize { xs, sm, md, lg }
+
+// https://material.io/design/layout/responsive-layout-grid.html#breakpoints
+class ScreenSizeBreakpoints {
+  static double xs = 0;
+  static double sm = 600;
+  static double md = 1240;
+  static double lg = 1440;
+}
+
 extension ContextExtensions on BuildContext {
-  bool get darkMode => Theme.of(this).brightness == Brightness.dark;
+  Brightness get brightness => Theme.of(this).brightness;
+
+  /// Check if brightness is Brightness.dark
+  bool get isDarkMode => brightness == Brightness.dark;
+
+  /// MediaQueryData for context
+  MediaQueryData get mq => MediaQuery.of(this);
+
+  /// Is device in landscape mode.
+  bool get isLandscape => mq.orientation == Orientation.landscape;
+
+  /// Is device in portrait mode.
+  bool get isPortrait => mq.orientation == Orientation.portrait;
+
+  /// Screen width
+  double get screenWidth => mq.size.width;
+
+  /// Screen height
+  double get screenHeight => mq.size.height;
+
+  /// Returns [ScreenSize] based on Material breakpoints
+  ScreenSize get screenSize {
+    return screenWidth >= ScreenSizeBreakpoints.lg
+        ? ScreenSize.lg
+        : screenWidth >= ScreenSizeBreakpoints.md
+            ? ScreenSize.md
+            : screenWidth >= ScreenSizeBreakpoints.sm
+                ? ScreenSize.sm
+                : ScreenSize.xs;
+  }
 }