+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Added screen size dynamic attribute.
+- Added device orientation dynamic attribute.
+
 ## 0.0.1
 
 - Initial release.